Zach ran into another problem with the T1 power series. Fixed compatibility problem with old odnp exp software 'jf_dnp' and 'rb_dnp'. This is kind of a screwy fix but should hold.
A better fix to this issue with figuring out if the first T1 experiment
was ran at full attenuation would be to accept or reject the first power
based on the minimum experiment time. Note that this would rely on the
assumption that the T1 experiment is longer than 1 minute, the time it
takes for the amplifier to warm up.